Room Specialization
Added a locked room and gave rooms requirements to enter that can be
specified. Added some future commands. Changed when the NPC moves
(currently bypasses locked rooms).

+++ Item.java
@@ -16,7 +16,7 @@ public class Item
     /**
      * Constructor for objects of class Item
      */
-    public Item(String name, float weight, boolean canTake)
+    public Item(String name, double weight, boolean canTake)
     {
         this.name = name;
         this.weight = weight;
